claude_multimodal/requirements.txt
jsonpickle==2.1.0
langchain==0.1.9
langchain-anthropic==0.1.3
langchain-community==0.0.24
langchain-core==0.1.29
langchain-text-splitters==0.0.1
langsmith==0.1.5
openai==1.12.0
opencv_python==4.9.0.80
pydantic<2.0
python-dotenv==1.0.1
ratelimit==2.2.1
requests<2.30.0
tqdm==4.66.2o